+++ app/src/androidTest/java/com/damienwesterman/defensedrill/domain/DrillGeneratorTest.java
@@ -9,7 +9,7 @@
  *                            *
  \****************************/
 
-package com.damienwesterman.defensedrill.util;
+package com.damienwesterman.defensedrill.domain;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -24,7 +24,6 @@ import androidx.test.platform.app.InstrumentationRegistry;
 
 import com.damienwesterman.defensedrill.data.local.Drill;
 import com.damienwesterman.defensedrill.data.local.DrillRepository;
-import com.damienwesterman.defensedrill.domain.DrillGenerator;
 
 import org.junit.Before;
 import org.junit.Test;
