+++ app/src/main/java/com/damienwesterman/defensedrill/ui/activity/CategorySelectActivity.java
@@ -44,7 +44,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.damienwesterman.defensedrill.R;
 import com.damienwesterman.defensedrill.ui.adapter.AbstractCategoryAdapter;
 import com.damienwesterman.defensedrill.ui.viewmodel.CategoryViewModel;
-import com.damienwesterman.defensedrill.util.Constants;
+import com.damienwesterman.defensedrill.common.Constants;
 
 import dagger.hilt.android.AndroidEntryPoint;
