+++ app/src/main/java/com/damienwesterman/defensedrill/ui/viewmodel/CategoryViewModel.java
@@ -37,7 +37,7 @@ import androidx.lifecycle.MutableLiveData;
 import com.damienwesterman.defensedrill.data.local.AbstractCategoryEntity;
 import com.damienwesterman.defensedrill.data.local.CategoryEntity;
 import com.damienwesterman.defensedrill.data.local.DrillRepository;
-import com.damienwesterman.defensedrill.ui.util.OperationCompleteCallback;
+import com.damienwesterman.defensedrill.common.OperationCompleteCallback;
 
 import java.util.ArrayList;
 import java.util.List;
