+++ app/src/main/java/com/damienwesterman/defensedrill/ui/viewmodel/SimulatedAttackSettingsViewModel.java
@@ -41,8 +41,8 @@ import com.damienwesterman.defensedrill.data.local.Drill;
 import com.damienwesterman.defensedrill.data.local.DrillRepository;
 import com.damienwesterman.defensedrill.data.local.SimulatedAttackRepo;
 import com.damienwesterman.defensedrill.data.local.WeeklyHourPolicyEntity;
-import com.damienwesterman.defensedrill.ui.util.OperationCompleteCallback;
-import com.damienwesterman.defensedrill.util.Constants;
+import com.damienwesterman.defensedrill.common.OperationCompleteCallback;
+import com.damienwesterman.defensedrill.common.Constants;
 
 import java.util.ArrayList;
 import java.util.HashMap;
