+++ app/src/main/res/layout/activity_category_select.xml
@@ -30,7 +30,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
 
-    <com.damienwesterman.defensedrill.ui.util.TitleDescCard
+    <com.damienwesterman.defensedrill.ui.view.TitleDescCard
         android:id="@+id/randomCategoryCard"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
