+++ app/src/main/res/layout/activity_web_drill_options.xml
@@ -27,7 +27,7 @@
         android:orientation="vertical"
         tools:context=".ui.activity.HomeActivity" >
 
-        <com.damienwesterman.defensedrill.ui.util.TitleDescCard
+        <com.damienwesterman.defensedrill.ui.view.TitleDescCard
             android:id="@+id/downloadFromDatabaseCard"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -38,7 +38,7 @@
             app:description="@string/download_from_database_description"
             android:onClick="onCardClick"/>
 
-        <com.damienwesterman.defensedrill.ui.util.TitleDescCard
+        <com.damienwesterman.defensedrill.ui.view.TitleDescCard
             android:id="@+id/loginCard"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -49,7 +49,7 @@
             app:description="@string/login_description"
             android:onClick="onCardClick"/>
 
-        <com.damienwesterman.defensedrill.ui.util.TitleDescCard
+        <com.damienwesterman.defensedrill.ui.view.TitleDescCard
             android:id="@+id/logoutCard"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
