+++ app/src/main/res/layout/layout_card_item.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
 
-    <com.damienwesterman.defensedrill.ui.util.TitleDescCard
+    <com.damienwesterman.defensedrill.ui.view.TitleDescCard
         android:id="@+id/cardViewHolder"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
